refactor: コードを整理
